requirements.txt
numpy
matplotlib
torch
tensorflow
torchvision
scikit-learn
keras
ensure
python-box
pyyaml
opencv-python
